MAINT: Pass float to np.full explicitly.

+++ tests/pipeline/test_technical.py
@@ -168,7 +168,7 @@ class TestFastStochasticOscillator(ZiplineTestCase):
         fso.compute(today, assets, out, closes, lows, highs)
 
         # Expected %K
-        assert_equal(out, np.full((3,), 200))
+        assert_equal(out, np.full((3,), 200, dtype=np.float64))
 
     @parameter_space(seed=range(5))
     def test_fso_expected_with_talib(self, seed):
